requirements.txt
tensorflow==2.20.0
numpy==2.3.5
scipy==1.16.3
pandas==2.3.3
pillow==12.0.0
streamlit==1.52.1
scikit-learn==1.8.0
matplotlib==3.10.8
